requirements.txt
asgiref==3.7.2
boto3==1.34.31
botocore==1.34.31
Django==5.0.1
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.14.0
jmespath==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.4
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.7
